added options to only show when wearing a full hunter armor
