+++ src/main/java/at/hannibal2/skyhanni/features/fishing/FishingAPI.kt
@@ -11,6 +11,7 @@ import at.hannibal2.skyhanni.features.fishing.trophy.TrophyFishManager
 import at.hannibal2.skyhanni.features.fishing.trophy.TrophyFishManager.getFilletValue
 import at.hannibal2.skyhanni.features.fishing.trophy.TrophyRarity
 import at.hannibal2.skyhanni.utils.BlockUtils.getBlockAt
+import at.hannibal2.skyhanni.utils.InventoryUtils
 import at.hannibal2.skyhanni.utils.ItemCategory
 import at.hannibal2.skyhanni.utils.ItemUtils.getInternalName
 import at.hannibal2.skyhanni.utils.ItemUtils.getItemCategoryOrNull
@@ -18,7 +19,9 @@ import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.LorenzVec
 import at.hannibal2.skyhanni.utils.NEUInternalName
 import at.hannibal2.skyhanni.utils.SimpleTimeMark
+import at.hannibal2.skyhanni.utils.StringUtils.matches
 import at.hannibal2.skyhanni.utils.getLorenzVec
+import at.hannibal2.skyhanni.utils.repopatterns.RepoPattern
 import net.minecraft.client.Minecraft
 import net.minecraft.entity.item.EntityArmorStand
 import net.minecraft.entity.projectile.EntityFishHook
@@ -28,6 +31,11 @@ import net.minecraftforge.event.entity.EntityJoinWorldEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 object FishingAPI {
+
+    private val trophyArmorNames by RepoPattern.pattern(
+        "fishing.trophyfishing.armor",
+        "(BRONZE|SILVER|GOLD|DIAMOND)_HUNTER_(HELMET|CHESTPLATE|LEGGINGS|BOOTS)"
+    )
 
     val lavaBlocks = listOf(Blocks.lava, Blocks.flowing_lava)
     private val waterBlocks = listOf(Blocks.water, Blocks.flowing_water)
@@ -42,6 +50,8 @@ object FishingAPI {
 
     var bobber: EntityFishHook? = null
     var bobberHasTouchedWater = false
+
+    var wearingTrophyArmor = false
 
     @SubscribeEvent
     fun onJoinWorld(event: EntityJoinWorldEvent) {
@@ -69,6 +79,11 @@ object FishingAPI {
     @SubscribeEvent
     fun onTick(event: LorenzTickEvent) {
         if (!LorenzUtils.inSkyBlock) return
+
+        if (event.isMod(5)) {
+            wearingTrophyArmor = isWearingTrophyArmor()
+        }
+
         val bobber = bobber ?: return
         if (bobber.isDead) {
             resetBobber()
@@ -144,5 +159,9 @@ object FishingAPI {
             return 2
         }
         return 1
+    }
+
+    private fun isWearingTrophyArmor(): Boolean = InventoryUtils.getArmor().all {
+        trophyArmorNames.matches(it?.getInternalName()?.asString())
     }
 }
